+++ source/tower/Tower.h
@@ -4,9 +4,9 @@
 #include "../variables/Variables.h"
 
 class Tower {
-    int dmg{};
+    double dmg{};
     int range{};
-    int cost{};
+    double cost{};
     int type{};
     SDL_FRect rect = {};
     SDL_Color color = {};
@@ -16,11 +16,11 @@ class Tower {
 
 public:
     Tower();
-    Tower(int dmg, int range, int cost, int type, SDL_FRect rect, SDL_Color color, SDL_Color hoverColor);
+    Tower(double dmg, int range, double cost, int type, SDL_FRect rect, SDL_Color color, SDL_Color hoverColor);
 
-    int getDmg();
+    double getDmg();
     int getRange();
-    int getCost();
+    double getCost();
     int getType();
     SDL_FRect& getRect();
     SDL_Color getColor();
@@ -28,9 +28,9 @@ public:
     bool getIsHovered();
     bool getIsVisible();
 
-    void setDmg(int dmg);
+    void setDmg(double dmg);
     void setRange(int range);
-    void setCost(int cost);
+    void setCost(double cost);
     void setType(int type);
     void setIsHovered(bool isHovered);
     void setIsVisible(bool isVisible);
@@ -39,7 +39,7 @@ public:
 
 };
 
-void CreateNewTower(int dmg, int range, int cost, int type, SDL_FRect _rect, SDL_Color _color, SDL_Color _hoverColor, int pos);
+void CreateNewTower(double dmg, int range, double cost, int type, SDL_FRect _rect, SDL_Color _color, SDL_Color _hoverColor, int pos);
 
 void renderTower(SDL_Renderer* renderer, Tower& tower);
